Inventory manager now loads items from DB on startup
